Streamagg3

+++ oups/collection.py
@@ -173,8 +173,9 @@ class ParquetSet:
             raise TypeError("data should be a pandas or vaex dataframe.")
         dirpath = os_path.join(self._basepath, key.to_path)
         write(dirpath=dirpath, data=data, md_key=key, **kwargs)
-        # If no trouble from writing, add key.
-        self._keys.add(key)
+        if key not in self:
+            # If no trouble from writing, add key.
+            self._keys.add(key)
         return
 
     def __setitem__(
@@ -232,14 +233,15 @@ class ParquetSet:
             Key specifying the location where to write the data. It has to be
             an instance of the dataclass provided at ParquetSet instantiation.
         """
-        # Keep track of intermediate partition folders, in case one get
-        # empty.
-        basepath = self._basepath
-        dirpath = os_path.join(basepath, key.to_path)
-        rmtree(dirpath)
-        self._keys.remove(key)
-        # Remove possibly empty directories.
-        upper_dir = strip_path_tail(dirpath)
-        while (upper_dir != basepath) and (not listdir(upper_dir)):
-            rmdir(upper_dir)
-            upper_dir = strip_path_tail(upper_dir)
+        if key in self._keys:
+            # Keep track of intermediate partition folders, in case one get
+            # empty.
+            basepath = self._basepath
+            dirpath = os_path.join(basepath, key.to_path)
+            rmtree(dirpath)
+            self._keys.remove(key)
+            # Remove possibly empty directories.
+            upper_dir = strip_path_tail(dirpath)
+            while (upper_dir != basepath) and (not listdir(upper_dir)):
+                rmdir(upper_dir)
+                upper_dir = strip_path_tail(upper_dir)
